Updating to gradle 3.2.1
In order to update from 3.0.0, ned mess dependencies etc. I spent many hours figuring out how to do it. At the end I was forced to create a new project with the same classes, that used the new gradle (new gradle is required for functionality of Robolectric). So this should be behind me and I should now fully focus on locak android-framework tests. We'll see how I'll try to make an widget test